+++ dart_native/example/ios/Runner/RuntimeStub.m
@@ -176,6 +176,34 @@ API_AVAILABLE(ios(11.0)){
     return array;
 }
 
+- (NSMutableArray *)fooNSMutableArray:(NSMutableArray *)array {
+    DDLogInfo(@"%s %@", __FUNCTION__, array.description);
+    [array addObject:@"mutable!"];
+    return array;
+}
+
+- (NSDictionary *)fooNSDictionary:(NSDictionary *)dict {
+    DDLogInfo(@"%s %@", __FUNCTION__, dict.description);
+    return dict;
+}
+
+- (NSMutableDictionary *)fooNSMutableDictionary:(NSMutableDictionary *)dict {
+    DDLogInfo(@"%s %@", __FUNCTION__, dict.description);
+    dict[@"newKey"] = @"mutable!";
+    return dict;
+}
+
+- (NSSet *)fooNSSet:(NSSet *)set {
+    DDLogInfo(@"%s %@", __FUNCTION__, set.description);
+    return set;
+}
+
+- (NSMutableSet *)fooNSMutableSet:(NSMutableSet *)set {
+    DDLogInfo(@"%s %@", __FUNCTION__, set.description);
+    [set addObject:@"mutable!"];
+    return set;
+}
+
 typedef NSObject *(^BarBlock)(NSObject *a);
 
 - (BarBlock)fooBlock:(BarBlock)block {
@@ -252,6 +280,22 @@ typedef char *(^CStringRetBlock)(char *a);
 - (NSString *)fooNSString:(NSString *)str {
     DDLogInfo(@"%s arg: %@", __FUNCTION__, str);
     return @"test nsstring";
+}
+
+- (NSMutableString *)fooNSMutableString:(NSMutableString *)str {
+    [str appendString:@" mutable!"];
+    DDLogInfo(@"%s arg: %@", __FUNCTION__, str);
+    return str;
+}
+
+- (void)fooWithError:(out NSError **)error {
+    if (error) {
+        *error = [NSError errorWithDomain:@"com.dartnative.test" code:-1 userInfo:nil];
+    }
+}
+
+- (TestOptions)fooWithOptions:(TestOptions)options {
+    return options;
 }
 
 @end