+++ dart_native/example/lib/ios/unit_test.dart
@@ -84,8 +84,23 @@ testIOS(RuntimeStub stub, DelegateStub delegate) {
       stub.fooCGAffineTransform(CGAffineTransform(6, 5, 4, 3, 2, 1));
   print('fooCGAffineTransform result:$transform');
 
-  NSArray array = stub.fooNSArray([1, 2.345, 'I\'m String', rect]);
-  print(array);
+  List list = stub.fooNSArray([1, 2.345, 'I\'m String', rect]);
+  print('NSArray to List: $list');
+
+  list = stub.fooNSMutableArray([1, 2.345, 'I\'m String', rect]);
+  print('NSMutableArray to List: $list');
+
+  Map map = stub.fooNSDictionary({'foo': 'bar'});
+  print('NSDictionary to Map: $map');
+
+  map = stub.fooNSMutableDictionary({'foo': 'bar'});
+  print('NSMutableDictionary to Map: $map');
+
+  Set set = stub.fooNSSet(Set.from([1, 2.345, 'I\'m String', rect]));
+  print('NSSet to Set: $set');
+
+  set = stub.fooNSMutableSet(Set.from([1, 2.345, 'I\'m String', rect]));
+  print('fooNSMutableSet to Set: $set');
 
   Block block = stub.fooBlock((NSString a) {
     print('hello block! ${a.toString()}');
@@ -94,7 +109,7 @@ testIOS(RuntimeStub stub, DelegateStub delegate) {
   resultObj = block.invoke([stub]);
   print('fooBlock result:$resultObj');
 
-  Block blockStret = stub.fooStretBlock((CGRect a) {
+  Block blockStret = stub.fooStretBlock((CGAffineTransform a) {
     print('hello block stret! ${a.toString()}');
     return CGAffineTransform(12, 0, 12, 0, 12, 0);
   });
@@ -112,13 +127,24 @@ testIOS(RuntimeStub stub, DelegateStub delegate) {
   stub.fooDelegate(delegate);
   stub.fooStructDelegate(delegate);
 
-  NSString resultNSString = stub.fooNSString('This is NSString');
+  String resultNSString = stub.fooNSString('This is NSString');
   print('fooNSString result:$resultNSString');
+
+  resultNSString = stub.fooNSMutableString('This is NSString');
+  print('fooNSMutableString result:$resultNSString');
+
+  NSObjectRef<NSError> ref = NSObjectRef<NSError>();
+  stub.fooWithError(ref);
+  print('fooWithError result:${ref.value.description}');
+
+  TestOptions options =
+      stub.fooWithOptions(TestOptions(TestOptionsOne | TestOptionsTwo));
+  print('fooWithOptions result:$options');
 
   NSObject currentThread = Class('NSThread')
       .perform(SEL('currentThread'), onQueue: DispatchQueue.global());
   NSObject description = currentThread.perform(SEL('description'));
-  String threadResult = NSString.fromPointer(description.pointer).value;
+  String threadResult = NSString.fromPointer(description.pointer).raw;
   print('currentThread: $threadResult');
 
   NSNotificationCenter.defaultCenter.addObserver(