+++ dart_native/lib/dart_native.dart
@@ -1,3 +1,6 @@
+library dart_native;
+
+export 'package:dart_native/src/common/annotation.dart';
 export 'package:dart_native/src/ios/dart_objc.dart';
 export 'package:dart_native/src/android/dart_java.dart';
 export 'package:dart_native/src/dart_native_root.dn.dart';