+++ dart_native/lib/src/ios/foundation/basic_type.dart
@@ -6,26 +6,34 @@ mixin _ToAlias {}
 
 class BOOL = NativeBox<bool> with _ToAlias;
 
-class char extends NativeBox<int> with _ToAlias {
-  char(int value) : super(value);
+class char extends NativeIntBox {
+  const char(int raw) : super(raw);
 
   @override
   String toString() {
-    return utf8.decode([value]);
+    return utf8.decode([raw]);
   }
 }
 
 class unsigned_char = char with _ToAlias;
-class short = NativeBox<int> with _ToAlias;
-class unsigned_short = NativeBox<int> with _ToAlias;
-class unsigned_int = NativeBox<int> with _ToAlias;
-class long = NativeBox<int> with _ToAlias;
-class unsigned_long = NativeBox<int> with _ToAlias;
-class long_long = NativeBox<int> with _ToAlias;
-class unsigned_long_long = NativeBox<int> with _ToAlias;
-class size_t = NativeBox<int> with _ToAlias;
-class NSInteger = NativeBox<int> with _ToAlias;
-class NSUInteger = NativeBox<int> with _ToAlias;
-class float = NativeBox<double> with _ToAlias;
-class CGFloat = NativeBox<double> with _ToAlias;
+class short = NativeIntBox with _ToAlias;
+class unsigned_short = NativeIntBox with _ToAlias;
+class unsigned_int = NativeIntBox with _ToAlias;
+class long = NativeIntBox with _ToAlias;
+class unsigned_long = NativeIntBox with _ToAlias;
+class long_long = NativeIntBox with _ToAlias;
+class unsigned_long_long = NativeIntBox with _ToAlias;
+class size_t = NativeIntBox with _ToAlias;
+class NSInteger = NativeIntBox with _ToAlias;
+class NSUInteger = NativeIntBox with _ToAlias;
+class float = NativeNumBox<double> with _ToAlias;
+class CGFloat = NativeNumBox<double> with _ToAlias;
 class CString = NativeBox<String> with _ToAlias;
+
+class NSEnum extends NativeIntBox {
+  const NSEnum(int raw) : super(raw);
+}
+
+class NSOptions extends NativeIntBox {
+  const NSOptions(int raw) : super(raw);
+}