+++ dart_native/lib/src/ios/foundation/collection/nsarray.dart
@@ -9,7 +9,7 @@ import 'package:dart_native_gen/dart_native_gen.dart';
 
 @native
 class NSArray extends NSSubclass<List> {
-  NSArray(List value) : super(value, _new) {
+  NSArray(List value, {InitSubclass init: _new}) : super(value, init) {
     value = List.of(value, growable: false);
   }
 
@@ -20,25 +20,7 @@ class NSArray extends NSSubclass<List> {
       id e = objectAtIndex(i);
       temp[i] = unboxingElementForDartCollection(e);
     }
-    value = temp;
-  }
-
-  static Pointer<Void> _new(dynamic value) {
-    if (value is List) {
-      List boxValues = value.map((e) {
-        return boxingElementForNativeCollection(e);
-      }).toList();
-      Pointer<Pointer<Void>> listPtr = allocate(count: boxValues.length);
-      for (var i = 0; i < boxValues.length; i++) {
-        listPtr.elementAt(i).value = boxValues[i].pointer;
-      }
-      NSObject result = Class('NSArray').perform(SEL('arrayWithObjects:count:'),
-          args: [listPtr, boxValues.length]);
-      free(listPtr);
-      return result.pointer;
-    } else {
-      throw 'Invalid param when initializing NSArray.';
-    }
+    raw = temp;
   }
 
   int get count => perform(SEL('count'));
@@ -48,6 +30,39 @@ class NSArray extends NSSubclass<List> {
   }
 }
 
+/// Only for type casting. It's unmodifiable.
+@native
+class NSMutableArray extends NSArray {
+  NSMutableArray(List value) : super(value, init: _mutableCopy) {
+    value = List.of(value, growable: true);
+  }
+
+  NSMutableArray.fromPointer(Pointer<Void> ptr) : super.fromPointer(ptr);
+
+  static Pointer<Void> _mutableCopy(dynamic value) {
+    return NSObject.fromPointer(_new(value)).mutableCopy().pointer;
+  }
+}
+
+Pointer<Void> _new(dynamic value) {
+  if (value is List) {
+    List boxValues = value.map((e) {
+      return boxingElementForNativeCollection(e);
+    }).toList();
+    Pointer<Pointer<Void>> listPtr = allocate(count: boxValues.length);
+    for (var i = 0; i < boxValues.length; i++) {
+      listPtr.elementAt(i).value = boxValues[i].pointer;
+    }
+    NSObject result = Class('NSArray').perform(SEL('arrayWithObjects:count:'),
+        args: [listPtr, boxValues.length]);
+    free(listPtr);
+    return result.pointer;
+  } else {
+    throw 'Invalid param when initializing NSArray.';
+  }
+}
+
 extension ConvertToNSArray on List {
   NSArray toNSArray() => NSArray(this);
+  NSMutableArray toNSMutableArray() => NSMutableArray(this);
 }