+++ dart_native/lib/src/ios/runtime.dart
@@ -1,5 +1,3 @@
-library dart_objc.runtime;
-
 export 'runtime/selector.dart';
 export 'runtime/class.dart';
 export 'runtime/protocol.dart';
@@ -7,3 +5,4 @@ export 'runtime/block.dart';
 export 'runtime/nsobject.dart';
 export 'runtime/functions.dart';
 export 'runtime/nsobject_protocol.dart';
+export 'runtime/nsobject_ref.dart';