+++ dart_native/lib/src/ios/runtime/id.dart
@@ -120,25 +120,37 @@ class id implements NSObjectProtocol {
   /// Returns a string that describes the contents of the receiver.
   String get description {
     NSObject result = perform(SEL('description'));
-    return NSString.fromPointer(result.pointer).value;
+    return NSString.fromPointer(result.pointer).raw;
   }
 
   /// Returns a string that describes the contents of the receiver for presentation in the debugger.
   String get debugDescription {
     NSObject result = perform(SEL('debugDescription'));
-    return NSString.fromPointer(result.pointer).value;
+    return NSString.fromPointer(result.pointer).raw;
   }
 
   /// Sends a specified message to the receiver and returns the result of the message.
   dynamic perform(SEL selector,
-      {List args, DispatchQueue onQueue, bool waitUntilDone = true}) {
-    return msgSend(this, selector, args, true, onQueue, waitUntilDone);
+      {List args,
+      DispatchQueue onQueue,
+      bool waitUntilDone = true,
+      bool decodeRetVal = true}) {
+    return msgSend(this.pointer, selector,
+        args: args,
+        onQueue: onQueue,
+        waitUntilDone: waitUntilDone,
+        decodeRetVal: decodeRetVal);
   }
 
   /// Returns a Boolean value that indicates whether the receiver does not descend from NSObject.
   bool isProxy() {
     return perform(SEL('isProxy'));
   }
+
+  /// BasicProtocol
+  /// All instances already implement these methods. Do nothing.
+  @override
+  register() {}
 
   @override
   String toString() {