+++ dart_native/lib/src/ios/runtime/nsobject.dart
@@ -17,8 +17,6 @@ class NSObject extends id {
     }
   }
 
-  NSObject.alloc([Class isa]) : super(_alloc(isa));
-
   NSObject init() {
     return perform(SEL('init'));
   }
@@ -33,14 +31,6 @@ class NSObject extends id {
     return NSObject.fromPointer(result.autorelease().pointer);
   }
 
-  static Pointer<Void> _alloc(Class isa) {
-    if (isa == null) {
-      isa = Class('NSObject');
-    }
-    NSObject result = isa.perform(SEL('alloc'));
-    return result.autorelease().pointer;
-  }
-
   static Pointer<Void> _new(Class isa) {
     if (isa == null) {
       isa = Class('NSObject');
@@ -48,6 +38,14 @@ class NSObject extends id {
     NSObject result = isa.perform(SEL('new'));
     return result.autorelease().pointer;
   }
+}
+
+Pointer<Void> alloc(Class isa) {
+  if (isa == null) {
+    isa = Class('NSObject');
+  }
+  NSObject result = isa.perform(SEL('alloc'));
+  return result.autorelease().pointer;
 }
 
 typedef dynamic ConvertorFromPointer(Pointer<Void> ptr);
@@ -59,11 +57,24 @@ void registerTypeConvertor(String type, ConvertorFromPointer convertor) {
 }
 
 dynamic convertFromPointer(String type, dynamic arg) {
+  Pointer<Void> ptr;
   if (arg is NSObject) {
-    ConvertorFromPointer convertor = _convertorCache[type];
-    if (convertor != null) {
-      return convertor(arg.pointer);
-    }
+    ptr = arg.pointer;
+  } else if (arg is Pointer) {
+    ptr = arg;
+  } else {
+    return arg;
+  }
+
+  if (ptr == nullptr) {
+    return arg;
+  }
+
+  ConvertorFromPointer convertor = _convertorCache[type];
+  if (convertor != null) {
+    return convertor(ptr);
+  } else if (arg is Pointer) {
+    return NSObject.fromPointer(arg);
   }
   return arg;
 }